+++ src/main/kotlin/io/github/realyusufismail/armourandtoolsmod/core/init/RecipeBookTypesInit.kt
@@ -22,4 +22,5 @@ import net.minecraft.world.inventory.RecipeBookType
 
 object RecipeBookTypesInit {
     val ARMOUR_CRAFTING: RecipeBookType = RecipeBookType.create("armour_crafting")
+    val TOOL_CRAFTING: RecipeBookType = RecipeBookType.create("tool_crafting")
 }