+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.4-1.0.0.beta.3] - 2023-4-21
+
+### Bug fixe where game crashed due to outdated patchouli version
+### Added new custom tool table
+
 ## [1.19.4-1.0.0.beta.2] - 2023-4-17
 
 ### Add all armour layers apart from rainbow which needs to be done.